Go back to using STL cos

+++ include/blur-decode.h
@@ -148,7 +148,7 @@ namespace blurhash
 			const float x_pi { std::numbers::pi_v< float > * static_cast< float >( x ) };
 			for ( int i = 0; i < components_x; ++i )
 				basics_x_precalc[ x ][ i ] =
-					cosf( ( x_pi * static_cast< float >( i ) ) / static_cast< float >( width ) );
+					std::cos( ( x_pi * static_cast< float >( i ) ) / static_cast< float >( width ) );
 		}
 
 		for ( int y = 0; y < height; ++y )
@@ -158,7 +158,7 @@ namespace blurhash
 
 			float basics_y[ components_y ];
 			for ( int j = 0; j < components_y; ++j )
-				basics_y[ j ] = cosf( ( y_pi * static_cast< float >( j ) ) / static_cast< float >( height ) );
+				basics_y[ j ] = std::cos( ( y_pi * static_cast< float >( j ) ) / static_cast< float >( height ) );
 
 			for ( int x = 0; x < width; ++x )
 			{
